chore(ffi): rename ffi package to match dir
